Separated noether_tpp classes into individual files
