+++ noether_tpp/include/noether_tpp/tool_path_modifiers/direction_of_travel_orientation_modifier.h
@@ -1,0 +1,15 @@
+#pragma once
+
+#include <noether_tpp/core/tool_path_modifier.h>
+
+namespace noether
+{
+/**
+ * @brief Aligns the x-axis of all waypoints with the direction of travel between adjacent waypoints
+ */
+struct DirectionOfTravelOrientationModifier : public OneTimeToolPathModifier
+{
+  ToolPaths modify(ToolPaths tool_paths) const override final;
+};
+
+}  // namespace noether
