+++ noether_tpp/include/noether_tpp/tool_path_modifiers/raster_organization_modifier.h
@@ -1,0 +1,20 @@
+#pragma once
+
+#include <noether_tpp/core/tool_path_modifier.h>
+
+namespace noether
+{
+/**
+ * @brief Organizes a tool path into a raster-style pattern
+ * @details  Waypoints are sorted within each tool path segment in ascending order by x-axis value. Tool path segments
+ * are sorted within each tool path in ascending order by the x-axis value of their first waypoint. Tool paths are
+ * sorted in ascending order by the y-axis value of the first waypoint in the first tool path segment.
+ */
+struct RasterOrganizationModifier : OneTimeToolPathModifier
+{
+  using OneTimeToolPathModifier::OneTimeToolPathModifier;
+
+  ToolPaths modify(ToolPaths) const override;
+};
+
+}  // namespace noether
