+++ noether_tpp/src/tool_path_modifiers/standard_edge_paths_organization_modifier.cpp
@@ -1,78 +1,10 @@
-#include <noether_tpp/tool_path_modifiers/organization_modifiers.h>
+#include <noether_tpp/tool_path_modifiers/standard_edge_paths_organization_modifier.h>
 #include <noether_tpp/utils.h>
 
 #include <numeric>
 
 namespace noether
 {
-ToolPaths RasterOrganizationModifier::modify(ToolPaths tool_paths) const
-{
-  ToolPathSegment first_seg = tool_paths.at(0).at(0);
-  Eigen::Isometry3d first_wp = first_seg.at(0);
-  Eigen::Isometry3d last_wp_first_seg = first_seg.at(first_seg.size() - 1);
-  const Eigen::Vector3d reference_segment_dir = estimateToolPathDirection(tool_paths.at(0));
-
-  for (ToolPath& tool_path : tool_paths)
-  {
-    // Sort the waypoints within each tool path segment by their distance along the reference direction
-    for (ToolPathSegment& segment : tool_path)
-    {
-      std::sort(segment.begin(),
-                segment.end(),
-                [segment, reference_segment_dir](const ToolPathWaypoint& a, const ToolPathWaypoint& b) {
-                  Eigen::Vector3d diff_from_start_b = b.translation() - segment.at(0).translation();
-                  Eigen::Vector3d diff_from_start_a = a.translation() - segment.at(0).translation();
-                  return diff_from_start_a.dot(reference_segment_dir) < diff_from_start_b.dot(reference_segment_dir);
-                });
-    }
-
-    // Sort the tool path segments within each tool path by the distance of their first waypoints along the reference
-    // direction
-    std::sort(tool_path.begin(),
-              tool_path.end(),
-              [tool_path, reference_segment_dir](const ToolPathSegment& a, const ToolPathSegment& b) {
-                Eigen::Vector3d diff_from_start_b = b.at(0).translation() - tool_path.at(0).at(0).translation();
-                Eigen::Vector3d diff_from_start_a = a.at(0).translation() - tool_path.at(0).at(0).translation();
-                return diff_from_start_a.dot(reference_segment_dir) < diff_from_start_b.dot(reference_segment_dir);
-              });
-  }
-
-  // Sort the tool paths by their distance along a vector that is perpendicular to the reference direction of travel
-  const Eigen::Vector3d reference_tool_paths_dir = estimateRasterDirection(tool_paths, reference_segment_dir);
-  std::sort(tool_paths.begin(),
-            tool_paths.end(),
-            [tool_paths, reference_tool_paths_dir, first_wp](const ToolPath& a, const ToolPath& b) {
-              Eigen::Vector3d diff_from_start_b = b.at(0).at(0).translation() - first_wp.translation();
-              Eigen::Vector3d diff_from_start_a = a.at(0).at(0).translation() - first_wp.translation();
-              return diff_from_start_a.dot(reference_tool_paths_dir) < diff_from_start_b.dot(reference_tool_paths_dir);
-            });
-
-  return tool_paths;
-}
-
-ToolPaths SnakeOrganizationModifier::modify(ToolPaths tool_paths) const
-{
-  // Create a raster
-  RasterOrganizationModifier raster;
-  tool_paths = raster.modify(tool_paths);
-
-  // Re-sort the tool paths at odd indices
-  for (std::size_t i = 1; i < tool_paths.size(); i += 2)
-  {
-    ToolPath& tool_path = tool_paths.at(i);
-
-    // Sort waypoints in each tool path segment by descending x-axis value (i.e. right to left)
-    for (ToolPathSegment& segment : tool_path)
-    {
-      std::reverse(segment.begin(), segment.end());
-    }
-
-    std::reverse(tool_path.begin(), tool_path.end());
-  }
-
-  return tool_paths;
-}
-
 /**
  * @brief Organizes the tool path segments within each tool path in the container such that the beginning of a tool path
  * is as close as possible to the end of the previous tool path (or a reference position, in the case of the first tool
